requirements.txt
fastapi
uvicorn
pyyaml
python-dotenv
openai
anthropic
sseclient-py
aiohttp